Detect plugin usages of ocamlbuild
The previous bulk-update work only tried to detect packages using
ocamlbuild as a build system; not packages distributing ocamlbuild
plugins, or any other library or executable linked with OCamlbuild. In
this latter (minority) use-case, adding a build-time dependency only
is insufficient.

Gabriel Radanne suggests detecting ocamlbuild plugins by looking at
the string "Ocamlbuild_plugin" appearing in source code. This base
idea needed to be refined quite a bit to remove a fair number of false
positives (see the details in the commit), but we are now able to
detect 65 packages that have a non-build-time dependency over
ocamlbuild (over the roughly 3000 that build-depend on it).

At the time of writing this patch, it produces no false positive on
the opam-repository (I manually looked at the list and refined
heuristics as needed); false negatives may be present.