opamCudf: provide machine-readable information on conflicts caused by cycles
For conflicts caused by cyclic actions, the opamCudf API would
previously only allow to observe the cycle with actions described as
strings for human pretty-printing.

We would like to write a script, using the opam API, that can react to
solution errors caused by cycles by patching the query and calling the
solver again. For this we need a machine-readable representation of
conflict cycles.

The present commit makes the minimal change to the opamCudf API to
make this possible:

- The internal representation of cycles is changed to take CUDF
  packages (the most general information) instead of strings.
- The `OpamCudf.cycle_conflicts` function has its type accordingly
  (this changes the module interface and may break library consumers).
- A new function `OpamCudf.conflict_cycles` is exposed to recover
  a cycle representation carrying "opam package" actions.

Remark: `conflict_cycles` is in the style of the current API on the
abstract `conflict` type, which is not very good: a conflict is
a variant/sum type in hiding (either an incompatibility in the query
or a cycle in the solution), and functions do something in one case
and silently return nothing in the other.

It would be nicer to expose a variant-sum type view of the conflict
type to users, but this should be discussed with OPAM maintainers, and
is not a requirement for our work.